Attatched the tableview to the class

+++ vaporweight/WeightSession.swift
@@ -12,7 +12,7 @@ import Foundation
 class WeightSession {
     var startTime : NSDate
     var endTime : NSDate?
-    var duration : Float32?
+    var duration : Int?
     
     init(startTime: NSDate) {
         self.startTime = startTime
